* Switch all BigIntegers to longs (ergo internals are int64, only reason javascript has to use bigintegers is because longs are only 58bit) * Start building the base for transaction building

+++ FleetSharp/Sigma/SigmaTypeCode.cs
@@ -90,7 +90,7 @@ namespace FleetSharp.Sigma
             return _createPrimitiveType(SigmaTypeCode.Long, null);
         }
 
-        public static dynamic SLong(int value)
+        public static dynamic SLong(long value)
         {
             return _createPrimitiveType(SigmaTypeCode.Long, value);
         }
